validate RelayState using HMAC-SHA1
According to the SAML Bindings specification "RelayState data MAY be
included with a SAML protocol message" transmitted with the
HTTP-Redirect binding. But "because the value is exposed to third-party
tampering, the entity SHOULD ensure that the value has not been tampered
with by using a checksum, a pseudo-random value, or similar means". In
the current implementation we create a "secret key" on starutp that is
later on used to create a keyed-hash message authentication code (HMAC)
using SHA-1 as the secure hash algorithm. What is currently hashed is the
"continue-url" information, to produce the following as relay state:
      <continue-url>:<hmac-sha1(continue-url, secret-key)>
On the arrival of the response the relay state data is again validated
and if the continue-url does not match its hashed value the response
is rejected.